requirements.txt
python-dotenv~=1.0.1
chardet~=5.2.0
tqdm~=4.66.4
requests~=2.32.3
pydantic~=2.7.4
pillow~=10.3.0
torch~=2.3.1
torchvision~=0.18.1
pytorch-lightning
timm
onnxruntime~=1.18.0
opencv-python~=4.10.0.84
transformers~=4.42.3
supervision==0.3.0
wandb~=0.17.4
pycocotools
coco_eval
fastapi~=0.111.0
uvicorn~=0.15.0
streamlit~=1.36.0
torchmetrics~=1.4.0.post0
firebase-admin~=6.5.0